+++ oci/layer/xattr_linux_test.go
@@ -1,0 +1,268 @@
+//go:build linux
+// +build linux
+
+// SPDX-License-Identifier: Apache-2.0
+/*
+ * umoci: Umoci Modifies Open Containers' Images
+ * Copyright (C) 2016-2025 SUSE LLC
+ * Copyright (C) 2020 Cisco Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package layer
+
+import (
+	"archive/tar"
+	"os"
+	"path/filepath"
+	"testing"
+
+	"github.com/stretchr/testify/assert"
+	"github.com/stretchr/testify/require"
+	"github.com/vbatts/go-mtree"
+
+	"github.com/opencontainers/umoci/pkg/fseval"
+	"github.com/opencontainers/umoci/pkg/system"
+)
+
+func getAllXattrs(t *testing.T, path string) map[string]string {
+	names, err := system.Llistxattr(path)
+	require.NoErrorf(t, err, "fetch all xattrs for %q", path)
+	xattrs := map[string]string{}
+	for _, name := range names {
+		value, err := system.Lgetxattr(path, name)
+		require.NoErrorf(t, err, "fetch xattr %q for %q", name, path)
+		xattrs[name] = string(value)
+	}
+	return xattrs
+}
+
+func TestUnpackGenerateRoundTrip_ComplexXattr_OverlayFS(t *testing.T) {
+	dir := t.TempDir()
+
+	testNeedsMknod(t)
+	testNeedsTrustedOverlayXattrs(t)
+
+	dentries := []struct {
+		tarDentry
+		remapXattrs map[string]string
+	}{
+		{
+			tarDentry{path: ".", ftype: tar.TypeDir, xattrs: map[string]string{
+				"trusted.overlay.opaque": "x",
+				"user.dummy.xattr":       "foobar",
+			}},
+			map[string]string{
+				"trusted.overlay.overlay.opaque": "x",
+				"user.dummy.xattr":               "foobar",
+			},
+		},
+		// Set a fake overlayfs xattr in the trusted.overlay namespace on a
+		// directory that contains entries. Since restoreMetadata() gets called
+		// on all parent directories when unpacking, this will cause
+		// restoreMetadata() to be run on the same directory multiple times.
+		// This lets us test that restoreMetadata will not re-apply the xattr
+		// escaping even after being called multiple times.
+		{
+			tarDentry{path: "foo/", ftype: tar.TypeDir, xattrs: map[string]string{
+				"trusted.overlay.fakexattr": "fakexattr",
+			}},
+			map[string]string{
+				"trusted.overlay.overlay.fakexattr": "fakexattr",
+			},
+		},
+		// Some subpaths with dummy overlayfs xattrs.
+		{
+			tarDentry{path: "foo/bar", ftype: tar.TypeReg, contents: "file", xattrs: map[string]string{
+				"trusted.overlay.whiteout": "foo",
+			}},
+			map[string]string{
+				"trusted.overlay.overlay.whiteout": "foo",
+			},
+		},
+		{
+			tarDentry{path: "foo/baz/", ftype: tar.TypeDir, xattrs: map[string]string{
+				"trusted.overlay.opaque": "y",
+			}},
+			map[string]string{
+				"trusted.overlay.overlay.opaque": "y",
+			},
+		},
+		// Several levels nested overlayfs xattrs.
+		{
+			tarDentry{path: "foo/extra-nesting/", ftype: tar.TypeDir, xattrs: map[string]string{
+				"trusted.overlay.overlay.opaque":                                "x",
+				"trusted.overlay.overlay.overlay.whiteout":                      "foobar",
+				"trusted.overlay.overlay.overlay.overlay.overlay.overlay.dummy": "dummy xattr",
+			}},
+			map[string]string{
+				"trusted.overlay.overlay.overlay.opaque":                                "x",
+				"trusted.overlay.overlay.overlay.overlay.whiteout":                      "foobar",
+				"trusted.overlay.overlay.overlay.overlay.overlay.overlay.overlay.dummy": "dummy xattr",
+			},
+		},
+		{
+			tarDentry{path: "foo/extra-nesting/reg", ftype: tar.TypeReg, contents: "reg", xattrs: map[string]string{
+				"trusted.overlay.overlay.overlay.overlay.overlay.dummy123": "dummy xattr 123",
+			}},
+			map[string]string{
+				"trusted.overlay.overlay.overlay.overlay.overlay.overlay.dummy123": "dummy xattr 123",
+			},
+		},
+	}
+
+	// With extraction using OverlayFSWhiteout we expect to get the remapped
+	// xattrs.
+	t.Run("OverlayFSWhiteout", func(t *testing.T) {
+		unpackOptions := UnpackOptions{
+			MapOptions: MapOptions{
+				Rootless: os.Geteuid() != 0,
+			},
+			WhiteoutMode: OverlayFSWhiteout,
+		}
+
+		te := NewTarExtractor(unpackOptions)
+
+		for _, de := range dentries {
+			hdr, rdr := tarFromDentry(de.tarDentry)
+			err := te.UnpackEntry(dir, hdr, rdr)
+			assert.NoErrorf(t, err, "UnpackEntry %s", hdr.Name)
+		}
+
+		for _, de := range dentries {
+			path := de.path
+			fullPath := filepath.Join(dir, path)
+
+			xattrs := getAllXattrs(t, fullPath)
+			assert.Equalf(t, de.remapXattrs, xattrs, "UnpackEntry(%q): expected to see %#v remapped properly", path, de.xattrs)
+
+			// Make sure none of the files are whiteouts.
+			_, isWo, err := isOverlayWhiteout(fullPath, fseval.Default)
+			require.NoErrorf(t, err, "isOverlayWhiteout(%q)", path)
+			assert.Falsef(t, isWo, "isOverlayWhiteout(%q): regular entries with overlayfs xattrs should not end up being unpacked with overlayfs whiteout xattrs", path)
+		}
+
+		t.Run("ToGenerateLayer", func(t *testing.T) {
+			// something reasonable
+			mtreeKeywords := []mtree.Keyword{
+				"size",
+				"type",
+				"uid",
+				"gid",
+				"mode",
+			}
+			deltas, err := mtree.Check(dir, nil, mtreeKeywords, fseval.Default)
+			assert.NoError(t, err, "mtree check")
+
+			reader, err := GenerateLayer(dir, deltas, &RepackOptions{
+				TranslateOverlayWhiteouts: unpackOptions.WhiteoutMode == OverlayFSWhiteout,
+			})
+			require.NoError(t, err, "generate layer")
+			defer reader.Close()
+
+			// We expect to get the exact same thing as the original archive
+			// entries in the new archive.
+			var wantDentries []tarDentry
+			for _, dentry := range dentries {
+				wantDentries = append(wantDentries, dentry.tarDentry)
+			}
+			checkLayerEntries(t, reader, wantDentries)
+		})
+
+		t.Run("ToGenerateInsertLayer", func(t *testing.T) {
+			reader := GenerateInsertLayer(dir, ".", false, &RepackOptions{
+				TranslateOverlayWhiteouts: unpackOptions.WhiteoutMode == OverlayFSWhiteout,
+			})
+			defer reader.Close()
+
+			// We expect to get the exact same thing as the original archive
+			// entries in the new archive.
+			var wantDentries []tarDentry
+			for _, dentry := range dentries {
+				wantDentries = append(wantDentries, dentry.tarDentry)
+			}
+			checkLayerEntries(t, reader, wantDentries)
+		})
+	})
+
+	// With extraction using OverlayFSWhiteout we expect to get the remapped
+	// xattrs.
+	t.Run("OCIStandardWhiteout", func(t *testing.T) {
+		unpackOptions := UnpackOptions{
+			MapOptions: MapOptions{
+				Rootless: os.Geteuid() != 0,
+			},
+			WhiteoutMode: OCIStandardWhiteout,
+		}
+
+		te := NewTarExtractor(unpackOptions)
+
+		for _, de := range dentries {
+			hdr, rdr := tarFromDentry(de.tarDentry)
+			err := te.UnpackEntry(dir, hdr, rdr)
+			assert.NoErrorf(t, err, "UnpackEntry %s", hdr.Name)
+		}
+
+		for _, de := range dentries {
+			path := de.tarDentry.path
+			fullPath := filepath.Join(dir, path)
+
+			xattrs := getAllXattrs(t, fullPath)
+			assert.Equalf(t, de.xattrs, xattrs, "UnpackEntry(%q): expected to see %#v not be remapped", path, de.xattrs)
+			assert.NotEqualf(t, de.remapXattrs, xattrs, "UnpackEntry(%q): expected to see %#v not be remapped", path, de.xattrs)
+		}
+
+		t.Run("ToGenerateLayer", func(t *testing.T) {
+			// something reasonable
+			mtreeKeywords := []mtree.Keyword{
+				"size",
+				"type",
+				"uid",
+				"gid",
+				"mode",
+			}
+			deltas, err := mtree.Check(dir, nil, mtreeKeywords, fseval.Default)
+			assert.NoError(t, err, "mtree check")
+
+			reader, err := GenerateLayer(dir, deltas, &RepackOptions{
+				TranslateOverlayWhiteouts: unpackOptions.WhiteoutMode == OverlayFSWhiteout,
+			})
+			require.NoError(t, err, "generate layer")
+			defer reader.Close()
+
+			// We expect to get the exact same thing as the original archive
+			// entries in the new archive.
+			var wantDentries []tarDentry
+			for _, dentry := range dentries {
+				wantDentries = append(wantDentries, dentry.tarDentry)
+			}
+			checkLayerEntries(t, reader, wantDentries)
+		})
+
+		t.Run("ToGenerateInsertLayer", func(t *testing.T) {
+			reader := GenerateInsertLayer(dir, ".", false, &RepackOptions{
+				TranslateOverlayWhiteouts: unpackOptions.WhiteoutMode == OverlayFSWhiteout,
+			})
+			defer reader.Close()
+
+			// We expect to get the exact same thing as the original archive
+			// entries in the new archive.
+			var wantDentries []tarDentry
+			for _, dentry := range dentries {
+				wantDentries = append(wantDentries, dentry.tarDentry)
+			}
+			checkLayerEntries(t, reader, wantDentries)
+		})
+	})
+}